Bonus1 about binary search tree
Input consists of several test cases. For each test case, the first line
contains two positive integers N (<=10) and L, which are the total
number of integers in an insertion sequence and the number of sequences
to be tested, respectively. The second line contains N positive
integers, separated by a space, which are the initially
inserted numbers. Then L lines follow, each contains a sequence of N
integers to be checked.